+++ packages/docs/.vitepress/theme/index.js
@@ -1,4 +1,5 @@
 import DefaultTheme from 'vitepress/theme';
+import TwoslashFloatingVue from '@shikijs/vitepress-twoslash/client';
 import Badge from './components/Badge.vue';
 import Badges from './components/Badges.vue';
 import PreviewPlayground from './components/PreviewPlayground.vue';
@@ -15,5 +16,6 @@ export default {
     app.component('IframePreviewPlayground', IframePreviewPlayground);
     app.component('TableOfContent', TableOfContent);
     app.component('Toc', TableOfContent);
+    app.use(TwoslashFloatingVue);
   },
 };