+++ packages/docs/components/AnchorScrollto/stories/app.js
@@ -1,0 +1,13 @@
+import { Base, createApp } from '@studiometa/js-toolkit';
+import { AnchorScrollTo } from '@studiometa/ui';
+
+class App extends Base {
+  static config = {
+    name: 'App',
+    components: {
+      'a[href^="#"]': AnchorScrollTo,
+    },
+  };
+}
+
+export default createApp(App);