+++ packages/playground/static/morphdom.js
@@ -1,0 +1,3 @@
+import morphdom from 'morphdom';
+
+export default morphdom;