+++ packages/tests/Draggable/Draggable.spec.ts
@@ -1,0 +1,26 @@
+import { describe, it, expect, vi } from 'vitest';
+import { Draggable } from '@studiometa/ui';
+
+function createEvent(type: string, data = {}, options = {}) {
+  const event = new Event(type, options);
+  Object.entries(data).forEach(([name, value]) => {
+    event[name] = value;
+  });
+
+  return event;
+}
+
+describe('The Draggable component', () => {
+  it('should move its root element', async () => {
+    const div = document.createElement('div');
+    const draggable = new Draggable(div);
+    vi.useFakeTimers();
+    draggable.$mount();
+    await vi.advanceTimersByTimeAsync(100);
+    div.dispatchEvent(createEvent('pointerdown', { x: 0, y: 0, button: 0 }));
+    document.dispatchEvent(createEvent('mousemove', { clientX: 10, clientY: 10 }));
+    await vi.advanceTimersByTimeAsync(16);
+    expect(div.style.transform).toBe('translate3d(10px, 10px, 0px)');
+    vi.useRealTimers();
+  });
+});