+++ packages/ui/Frame/FrameLoader.ts
@@ -1,0 +1,33 @@
+import type { BaseProps, BaseConfig } from '@studiometa/js-toolkit';
+import { Transition } from '../Transition/index.js';
+import type { TransitionProps } from '../decorators/withTransition.js';
+
+export interface FrameLoaderProps extends BaseProps {
+  $options: TransitionProps['$options'] & {
+    enterKeep: true;
+    leaveKeep: true;
+  };
+}
+
+/**
+ * Class.
+ */
+export class FrameLoader<T extends BaseProps = BaseProps> extends Transition<T & FrameLoaderProps> {
+  /**
+   * Config.
+   */
+  static config: BaseConfig = {
+    name: 'FrameLoader',
+  };
+
+  // @ts-expect-error this is fine.
+  get $options() {
+    const options = super.$options;
+
+    return {
+      ...options,
+      enterKeep: true,
+      leaveKeep: true,
+    };
+  }
+}