+++ packages/ui/Slider/SliderCount.ts
@@ -1,3 +1,4 @@
+import { domScheduler } from '@studiometa/js-toolkit/utils';
 import type { BaseProps, BaseConfig } from '@studiometa/js-toolkit';
 import { AbstractSliderChild } from './AbstractSliderChild.js';
 
@@ -23,10 +24,10 @@ export class SliderCount<T extends BaseProps = BaseProps> extends AbstractSlider
 
   /**
    * Update the current counter indicator.
-   * @param   {number} index The new active index.
-   * @returns {void}
    */
   update(index: number) {
-    this.$refs.current.innerHTML = `${index + 1}`;
+    domScheduler.write(() => {
+      this.$refs.current.textContent = `${index + 1}`;
+    });
   }
 }